fix email verification

+++ Core/Interfaces/IFoodRepository.cs
@@ -8,11 +8,10 @@ namespace Core.Interfaces
         public Task<Food> UpdateAsync(Food food);
         public Task<List<Food>> GetAllAsync(string? search = null, string? orderBy = null);
         public Task<bool> IsFoodRated(string appUserId, int foodId);
-        public Task AddFoodRate(string appUserId, int foodId, int rate);
+        public Task AddFoodRate(string appUserId, int foodId, double rate);
         public Task UpdateFoodRate(FoodRating foodRating);
-        public Task UpdateFoodsRate(int foodId, int rate);
+        public Task UpdateFoodsRate(int foodId, double rate);
         public Task<List<Food>> GetFoodsByListOfIdsAsync(List<int> foodsId);
         public Task<int> GetFoodIdByName(string name);
-
     }
 }
